+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 0.2 (planned)
 +++++++++++++
 
+- Purge implicit form of Connection context manager.
 - Purge ``use_connection`` as it already deprecated in the RQ package.
 
 0.1 (2016-01-03)